refactor: modularize Express.js server structure (Closes #4)
