+++ server/Models/index.models.js
@@ -1,0 +1,9 @@
+const Details = require("./Booking/Booking.schema");
+const Registration = require("./User/User.schema");
+
+const Models = {
+    UserSchema: Registration,
+    BookingSchema: Details
+}
+
+module.exports = Models;
